fix(project): use proper relative mouse coordinates and cleanup small pieces of code and examples

+++ examples/sprite.rs
@@ -22,7 +22,7 @@ struct GameState {
 
 impl PixelGame for GameState {
     // Update loop exposing input events we can handle, this is where you would handle the game logic
-    fn update(&mut self, input: &Input, mouse_pos: Option<Vec2<usize>>, dt: f64) -> bool {
+    fn update(&mut self, input: &Input, mouse_pos: Option<Vec2<f64>>, dt: f64) -> bool {
         // If the left mouse button is pressed add a new sprite
         if let Some(mouse_pos) = mouse_pos {
             if input.mouse_pressed(MouseButton::Left) {
